Change base URL to use the HTTP Host header.
Use the HTTP 'Host' header when available, to make sure the lineup is
generated on whatever hostname the client contacted the server on.
Fall back to the old logic if the 'Host' header is missing (possible on
old\janky HTTP 1.0 clients, where it wasn't a required field). This
should help resolve the confusion in setting addresses in configs.